requirements.txt
python-telegram-bot==20.7
```

---

### Файл 3: `Procfile`
```
worker: python bot.py
```

---

### Файл 4: `runtime.txt`
```
python-3.11.0
```

---

### Файл 5: `.env` (для локального тестування)
```
BOT_TOKEN=8406711319:AAFbS0fNyOyRHdo_Ub3zZXU92E5I-6gqZmU
WEB_APP_URL=https://storied-daffodil-9cfeac.netlify.app
```

---

## ✅ Структура папки має бути:
```
ad-traffic-bot/
├── bot.py
├── requirements.txt
├── Procfile
├── runtime.txt
└── .env